Update the dependency ranges for the dropbox-sdk
Since this library isn't very updated I loosened the constraints
considerably. It is likely to be easier to add upper bounds here only
when a conflict occurs rather than maintain conservative bounds for this
package. Most of these packages are widely used with well maintained build-deps.

Also removed tls-extra which is no longer required with the newer
packages.